+++ cpp stuff/SquareMatrixF.h
@@ -9,19 +9,19 @@ class SquareMatrixF{
         SquareMatrixF(int size);
         SquareMatrixF(SquareMatrixF& source);
         ~SquareMatrixF();
-        void setElement(int x, int y, float value);
-        float getElement(int x, int y);
+        void setElement(int x, int y, double value);
+        double getElement(int x, int y);
         int getSize();
         void order(int* order);
-        Image toImage(float maxDist, int offset = 0, bool flipVertical = true, bool flipHorizontal = false);
-        Image toImage(float maxDist, ColorScheme* colorScheme, int offset = 0, bool flipVertical = true, bool flipHorizontal = false);
+        Image toImage(double maxDist, int offset = 0, bool flipVertical = true, bool flipHorizontal = false);
+        Image toImage(double maxDist, ColorScheme* colorScheme, int offset = 0, bool flipVertical = true, bool flipHorizontal = false);
         void toTspFullMatrix(char* fileNameWithPath, char* tspName = "tsplib from matrix", char* comment = "");
     private:
         void flipVertical();
         void flipHorizontal();
         void offset(int offset);
         int _size;
-        float* matrixBuffer;
+        double* matrixBuffer;
 };
 
 #endif
